+++ Fyber/src/main/java/com/mopub/mobileads/FyberInterstitial.java
@@ -1,18 +1,3 @@
-/* Copyright 2020 Fyber N.V.
-
- Licensed under the Apache License, Version 2.0 (the "License");
- you may not use this file except in compliance with the License.
- You may obtain a copy of the License at
-
- http://www.apache.org/licenses/LICENSE-2.0
-
- Unless required by applicable law or agreed to in writing, software
- distributed under the License is distributed on an "AS IS" BASIS,
- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- See the License for the specific language governing permissions and
- limitations under the License
- */
-
 package com.mopub.mobileads;
 
 import android.app.Activity;
@@ -38,31 +23,34 @@ import com.mopub.common.LifecycleListener;
 import com.mopub.common.MoPub;
 import com.mopub.common.Preconditions;
 import com.mopub.common.logging.MoPubLog;
-import com.mopub.mobileads.AdData;
-import com.mopub.mobileads.BaseAd;
-import com.mopub.mobileads.MoPubErrorCode;
 
 import java.util.Map;
 
+import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.CLICKED;
 import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.CUSTOM;
+import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.LOAD_ATTEMPTED;
+import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.LOAD_FAILED;
+import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.LOAD_SUCCESS;
+import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.SHOW_ATTEMPTED;
+import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.SHOW_FAILED;
+import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.SHOW_SUCCESS;
+import static com.mopub.mobileads.MoPubErrorCode.FULLSCREEN_LOAD_ERROR;
+import static com.mopub.mobileads.MoPubErrorCode.INLINE_LOAD_ERROR;
 
-/**
- * Implements Fyber's interstitial Mopub's custom event class
- */
 public class FyberInterstitial extends BaseAd {
-  // Mopub log tag definition
+
   private final static String LOG_TAG = "FyberInterstitialForMopub";
+  private static final String ADAPTER_NAME = FyberInterstitial.class.getSimpleName();
 
-  // Members
-  /**
-   * Fyber's interstitial ad object
-   */
+  private final FyberAdapterConfiguration mFyberAdapterConfiguration;
+
+  public FyberInterstitial() {
+    mFyberAdapterConfiguration = new FyberAdapterConfiguration();
+  }
+
   InneractiveAdSpot mInterstitialSpot;
-
   String mSpotId;
-  /**
-   * Context for showing the Ad
-   */
+
   @Nullable
   Context mContext;
 
@@ -88,20 +76,20 @@ public class FyberInterstitial extends BaseAd {
     Preconditions.checkNotNull(context);
     Preconditions.checkNotNull(adData);
 
-    log("load interstitial requested");
+    MoPubLog.log(getAdNetworkId(), LOAD_ATTEMPTED, ADAPTER_NAME);
 
     mContext = context;
-
     setAutomaticImpressionAndClickTracking(false);
 
     final Map<String, String> extras = adData.getExtras();
 
-    // Set variables from MoPub console.
     final String appId = extras == null ? null : extras.get(FyberMoPubMediationDefs.REMOTE_KEY_APP_ID);
     final String spotId = extras == null ? null : extras.get(FyberMoPubMediationDefs.REMOTE_KEY_SPOT_ID);
 
     if (TextUtils.isEmpty(spotId)) {
-      log("No spotID defined for ad unit. Cannot load interstitial");
+      MoPubLog.log(getAdNetworkId(), LOAD_FAILED, ADAPTER_NAME,
+              MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR.getIntCode(),
+              MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR);
       if (mLoadListener != null) {
         mLoadListener.onAdLoadFailed(MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR);
       }
@@ -110,7 +98,6 @@ public class FyberInterstitial extends BaseAd {
 
     mSpotId = spotId;
 
-    // If we've received an appId for this unit, try initializing the Fyber Marketplace SDK, if it was not already initialized
     if (!TextUtils.isEmpty(appId)) {
       FyberAdapterConfiguration.initializeFyberMarketplace(context, appId, extras.containsKey(
               FyberMoPubMediationDefs.REMOTE_KEY_DEBUG),
@@ -122,6 +109,9 @@ public class FyberInterstitial extends BaseAd {
                   if (status == OnFyberMarketplaceInitializedListener.FyberInitStatus.SUCCESSFULLY || status == OnFyberMarketplaceInitializedListener.FyberInitStatus.FAILED) {
                     requestInterstitial(context, spotId, extras);
                   } else if (mLoadListener != null) {
+                    MoPubLog.log(getAdNetworkId(), LOAD_FAILED, ADAPTER_NAME,
+                            MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR.getIntCode(),
+                            MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR);
                     mLoadListener.onAdLoadFailed(MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR);
                   }
                 }
@@ -129,90 +119,69 @@ public class FyberInterstitial extends BaseAd {
     } else if (InneractiveAdManager.wasInitialized()) {
       requestInterstitial(context, spotId, extras);
     } else if (mLoadListener != null) {
+      MoPubLog.log(getAdNetworkId(), LOAD_FAILED, ADAPTER_NAME,
+              MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR.getIntCode(),
+              MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR);
       mLoadListener.onAdLoadFailed(MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR);
     }
+
+    mFyberAdapterConfiguration.setCachedInitializationParameters(context, extras);
   }
 
-  /**
-   * Called by the Mopub infra-structure in order for the plugin to start showing a Fyber Marketplace interstitial
-   */
   @Override
   public void show() {
-    log("show interstitial called");
-    // check if the ad is ready
-    if (mInterstitialSpot != null && mInterstitialSpot.isReady()) {
+    MoPubLog.log(getAdNetworkId(), SHOW_ATTEMPTED, ADAPTER_NAME);
 
+    if (mInterstitialSpot != null && mInterstitialSpot.isReady()) {
       InneractiveFullscreenUnitController fullscreenUnitController = (InneractiveFullscreenUnitController) mInterstitialSpot
               .getSelectedUnitController();
       fullscreenUnitController.setEventsListener(new InneractiveFullscreenAdEventsListener() {
 
-        /**
-         * Called by Fyber Marketplace when an interstitial ad activity is closed
-         * @param adSpot Spot object
-         */
         @Override
         public void onAdDismissed(InneractiveAdSpot adSpot) {
-          log("onAdDismissed");
-          if (mInteractionListener != null) {
+            MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME, "Fyber interstitial is dismissed");
+
+            if (mInteractionListener != null) {
             mInteractionListener.onAdDismissed();
           }
         }
 
-        /**
-         * Called by Fyber Marketplace when an interstitial ad activity is shown
-         * @param adSpot Spot object
-         */
         @Override
         public void onAdImpression(InneractiveAdSpot adSpot) {
-          log("onAdImpression");
+          MoPubLog.log(getAdNetworkId(), SHOW_SUCCESS, ADAPTER_NAME);
+
           if (mInteractionListener != null) {
             mInteractionListener.onAdShown();
             mInteractionListener.onAdImpression();
           }
         }
 
-        /**
-         * Called by Fyber Marketplace when an interstitial ad is clicked
-         * @param adSpot Spot object
-         */
         @Override
         public void onAdClicked(InneractiveAdSpot adSpot) {
-          log("onAdClicked");
+          MoPubLog.log(getAdNetworkId(), CLICKED, ADAPTER_NAME);
+
           if (mInteractionListener != null) {
             mInteractionListener.onAdClicked();
           }
         }
 
-        /**
-         * Called by Fyber Marketplace when an interstitial ad opened an external application
-         * @param adSpot Spot object
-         */
         @Override
         public void onAdWillOpenExternalApp(InneractiveAdSpot adSpot) {
           log("onAdWillOpenExternalApp");
           // Don't call the onLeaveApplication() API since it causes a false Click event on MoPub
         }
 
-        /**
-         * Called when an ad has entered an error state, this will only happen when the ad is being shown
-         * @param adSpot the relevant ad spot
-         */
         @Override
         public void onAdEnteredErrorState(InneractiveAdSpot adSpot, AdDisplayError error) {
           log("onAdEnteredErrorState - " + error.getMessage());
         }
 
-        /**
-         * Called by Fyber Marketplace when Fyber Marketplace's internal browser, which was opened by this interstitial, was closed
-         * @param adSpot Spot object
-         */
         @Override
         public void onAdWillCloseInternalBrowser(InneractiveAdSpot adSpot) {
           log("onAdWillCloseInternalBrowser");
         }
       });
 
-      // Add video content controller, for controlling video ads
       InneractiveFullscreenVideoContentController videoContentController = new InneractiveFullscreenVideoContentController();
       videoContentController.setEventsListener(new VideoContentListener() {
         @Override
@@ -221,55 +190,36 @@ public class FyberInterstitial extends BaseAd {
                   + " position = " + positionInMsec);
         }
 
-        /**
-         * Called by Fyber Marketplace when an Interstitial video ad was played to the end
-         * <br>Can be used for incentive flow
-         * <br>Note: This event does not indicate that the interstitial was closed
-         */
         @Override
         public void onCompleted() {
-          log("Got video content completed event");
+          MoPubLog.log(CUSTOM, ADAPTER_NAME, "Video content completed event");
         }
 
         @Override
         public void onPlayerError() {
-          log("Got video content play error event");
+          MoPubLog.log(CUSTOM, ADAPTER_NAME, "Video content play error event");
         }
       });
 
-      // Now add the content controller to the unit controller
+
       fullscreenUnitController.addContentController(videoContentController);
 
       fullscreenUnitController.show((Activity) mContext);
     } else {
+      MoPubLog.log(getAdNetworkId(), SHOW_FAILED, ADAPTER_NAME);
       if (mInteractionListener != null) {
         mInteractionListener.onAdFailed(MoPubErrorCode.EXPIRED);
       }
-      log("The Interstitial ad is not ready yet.");
     }
   }
 
-
-  /**
-   * Called by Mopub when an ad should be destroyed
-   * <br>Destroy the underline Fyber Marketplace ad
-   */
-  @Override
   protected void onInvalidate() {
-    log("onInvalidate called by Mopub");
-    // We do the cleanup on the event of loadInterstitial.
     if (mInterstitialSpot != null) {
       mInterstitialSpot.destroy();
       mInterstitialSpot = null;
     }
   }
 
-  /**
-   * requests an interstitial ad from Fyber Marketplace
-   * @param context
-   * @param spotId
-   * @param localExtras
-   */
   private void requestInterstitial(final Context context, String spotId, Map<String, String> localExtras) {
     mContext = context;
 
@@ -280,7 +230,6 @@ public class FyberInterstitial extends BaseAd {
     }
 
     mInterstitialSpot = InneractiveAdSpotManager.get().createSpot();
-    // Set your mediation name and version
     mInterstitialSpot.setMediationName(InneractiveMediationName.MOPUB);
     mInterstitialSpot.setMediationVersion(MoPub.SDK_VERSION);
 
@@ -290,31 +239,21 @@ public class FyberInterstitial extends BaseAd {
     InneractiveAdRequest request = new InneractiveAdRequest(spotId);
     FyberAdapterConfiguration.updateRequestFromExtras(request, localExtras);
 
-    // Load ad
     mInterstitialSpot.setRequestListener(new InneractiveAdSpot.RequestListener() {
-
-      /**
-       * Called by Fyber Marketplace when an interstitial is ready for display
-       * @param adSpot Spot object
-       */
 
       @Override
       public void onInneractiveSuccessfulAdRequest(InneractiveAdSpot adSpot) {
-        log("on ad loaded successfully");
+        MoPubLog.log(getAdNetworkId(), LOAD_SUCCESS, ADAPTER_NAME);
+
         if (mLoadListener != null) {
           mLoadListener.onAdLoaded();
         }
       }
 
-      /**
-       * Called by Fyber Marketplace when an interstitial fails loading
-       * @param adSpot Spot object
-       * @param errorCode the failure's error.
-       */
       @Override
       public void onInneractiveFailedAdRequest(InneractiveAdSpot adSpot,
                                                InneractiveErrorCode errorCode) {
-        log("Failed loading interstitial with error: " + errorCode);
+
         if (mLoadListener != null) {
           if (errorCode == InneractiveErrorCode.CONNECTION_ERROR) {
             mLoadListener.onAdLoadFailed(MoPubErrorCode.NO_CONNECTION);
@@ -326,16 +265,14 @@ public class FyberInterstitial extends BaseAd {
             mLoadListener.onAdLoadFailed(MoPubErrorCode.SERVER_ERROR);
           }
         }
+        MoPubLog.log(getAdNetworkId(), LOAD_FAILED, ADAPTER_NAME, FULLSCREEN_LOAD_ERROR.getIntCode(),
+                errorCode);
       }
     });
 
     mInterstitialSpot.requestAd(request);
   }
 
-  /**
-   * MopubLog helper
-   * @param message
-   */
   private void log(String message) {
     MoPubLog.log(CUSTOM, LOG_TAG, message);
   }