+++ Fyber/CHANGELOG.md
@@ -2,5 +2,5 @@
 
   * 7.8.1.0
     * Initial commit
-    * Supports Banner, Interstitial and Rewarded ads.
-    * This version of the adapters has been certified with Fyber Marketplace 7.8.1 and MoPub 5.15.0.
+    * Add support for banner, interstitial and rewarded ads.
+    * This version of the adapters has been certified with Fyber Marketplace 7.8.1 and MoPub 5.16.2.